+++ src/datastructpy/__init__.py
@@ -4,6 +4,7 @@ from .base_comparer import BaseComparer, DefaultComparer
 from .exceptions import EmptyCollectionException
 from .hash_table import HashTable
 from .linked_list import LinkedList, LinkedListNode
+from .priority_queue import PriorityQueue
 from .queue import Queue
 from .stack import Stack
 
@@ -14,6 +15,7 @@ __all__ = [
     "HashTable",
     "LinkedList",
     "LinkedListNode",
+    "PriorityQueue",
     "Stack",
     "Queue",
 ]